+++ Simple Game Library/src/at/fhooe/mtd/sgl/graphics/screen/WindowedState.java
@@ -266,4 +266,12 @@ class WindowedState extends ScreenState {
         g.dispose();
         return fm;
     }
+
+    @Override
+    public Graphics2D getCurrentContext() {
+        if (g2d == null) {
+            throw new IllegalStateException("outside render cycle");
+        }
+        return g2d;
+    }
 }