+++ Simple Game Library/CHANGES.md
@@ -1,10 +1,13 @@
 # Version History
 
 ## Version 1.1.0
-Date: *2016-04-05*
+Date: *2016-04-13*
 
 - Added support to retrieve font metrics outside the rendering cycle.
-- Added readme file.
+- Added method `getCurrentContext()` to retrieve the current graphics context
+in situations when `beginUpdate()` has already been called and the returned context
+is not available.
+- Added *read-me* file.
 
 ## Version 1.0.0
 Date: *2016-03-30*